WeakMethod4 Added and Core changes performed
upon fully understanding that when we talk about an exception, and how
it should and should not be covered, we have updated our code to not
throw the exceptions, but instead, to let them happen to return them as
a value to the expected variable in the test cases files. Also,
WeakMethod4a and b were added, with full coverage of the branches and
the conditions.